requirements.txt
asgiref==3.7.2
bump2version==1.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.7
django-appconf==1.0.5
django-compressor==4.3.1
django-libsass==0.9
django-multiupload==0.6.1
djangorestframework==3.14
flake8==6.0.0
idna==3.4
joblib==1.2.0
libsass==0.22.0
nltk==3.8.1
numpy==1.25.0
oauthlib==3.2.2
osometweet==0.4
python-dotenv==1.0.0
rcssmin==1.1.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.65.0
twitterbloc @ git+https://github.com/anwala/bloc@cluster
typing_extensions==4.6.3
urllib3==1.26.6
